Created modules and components folders
