+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
-import { NotFoundComponent } from './not-found/not-found.component';
-import { CartComponent } from './cart/cart.component';
+import { CartComponent } from './components/cart/cart.component';
+import { NotFoundComponent } from './components/not-found/not-found.component';
 
 const routes: Routes = [
   {
     path: '',
-    loadChildren: "./shop.module#ShopModule"
+    loadChildren: "./modules/shop/shop.module#ShopModule"
   },
   { 
     path: 'cart',
