+++ src/app/components/cart/cart.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { CartService } from '../cart.service';
+import { CartService } from 'src/app/cart.service';
 
 
 @Component({
