+++ src/app/components/product-details/product-details.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 //import { products } from '../products';
-import { CartService } from '../cart.service';
+import { CartService } from 'src/app/cart.service';
 
 @Component({
   selector: 'app-product-details',
