+++ src/app/components/product-list/product-list.component.ts
@@ -1,0 +1,33 @@
+import { Component } from '@angular/core';
+import { CartService } from '../../cart.service';
+
+@Component({
+  selector: 'app-product-list',
+  templateUrl: './product-list.component.html',
+  styleUrls: ['./product-list.component.css']
+})
+export class ProductListComponent {
+  products;
+  constructor(private cartService: CartService){}
+  share() {
+    window.alert('The product has been shared!');
+  }
+
+  ngOnInit() {
+    this.products = this.cartService.getProductsList();
+  }
+
+  addToCart(product) {
+    window.alert('Your product has been added to the cart!');
+    this.cartService.addToCart(product);
+  }
+
+
+}
+
+
+/*
+Copyright Google LLC. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/
