+++ src/app/components/top-bar/top-bar.component.css
@@ -1,0 +1,7 @@
+
+
+/*
+Copyright Google LLC. All Rights Reserved.
+Use of this source code is governed by an MIT-style license that
+can be found in the LICENSE file at http://angular.io/license
+*/
