+++ src/app/modules/shop/shop.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ShopRoutingModule } from './shop-routing.module';
 
-import { CategoryComponent } from './category/category.component';
-import { ProductListComponent } from './product-list/product-list.component';
-import { ProductDetailsComponent } from './product-details/product-details.component';
+import { CategoryComponent } from '../../components/category/category.component';
+import { ProductListComponent } from '../../components/product-list/product-list.component';
+import { ProductDetailsComponent } from '../../components/product-details/product-details.component';
 import { MatCardModule } from '@angular/material/card';
