--- src/app/product-list/product-list.component.html
@@ -1,41 +1,0 @@
-<h2>Products</h2>
-<div class="row">
-<div *ngFor="let product of products | async" class="column">
-    <mat-card class="example-card">
-            <mat-card-header>
-                <mat-card-title>
-                    <h3>
-                        <a [title]="product.name + ' details'" [routerLink]="['/products', product.id]">
-                            {{ product.name }}
-                        </a>
-                    </h3>
-                </mat-card-title>
-                <mat-card-subtitle>Subtitle</mat-card-subtitle>
-            </mat-card-header>
-            <img mat-card-image src={{product.img}} class="img">
-            <mat-card-content>
-                <p *ngIf="product.description">
-                    Description: {{ product.description }}
-                </p>
-                <p *ngIf="product.price">
-                        Price: {{ product.price }}
-                    </p>
-            </mat-card-content>
-            <mat-card-actions>
-                <button (click)="share()" type="button" class="button fancy-button">
-                        <i class="material-icons">share</i> Share
-                </button>
-                <button (click)="addToCart(product)" class="button fancy-button"><i class="material-icons">add_shopping_cart</i>Buy</button>
-        </mat-card-actions>
-    </mat-card>
-</div>
-
-</div>
-
-
-
-<!-- 
-Copyright Google LLC. All Rights Reserved.
-Use of this source code is governed by an MIT-style license that
-can be found in the LICENSE file at http://angular.io/license
--->
