--- src/app/top-bar/top-bar.component.html
@@ -1,20 +1,0 @@
-<a [routerLink]="['/']">
-  <h1>My Store</h1>
-</a>
-<a [routerLink]="['/shop']">
-  <h1>Category</h1>
-</a>
-<a [routerLink]="['/about']">
-  <h1>About</h1>
-</a>
-<a [routerLink]="['/contact']">
-  <h1>Contact Us</h1>
-</a>
-<a [routerLink]="['/cart']" class="button fancy-button">
-  <i class="material-icons">shopping_cart</i>Checkout</a>
-
-<!-- 
-Copyright Google LLC. All Rights Reserved.
-Use of this source code is governed by an MIT-style license that
-can be found in the LICENSE file at http://angular.io/license
--->
